updates item with item_properties

+++ database/migrations/2020_11_26_145700_create_item_properties_table.php
@@ -14,7 +14,7 @@ class CreateItemPropertiesTable extends Migration
     public function up()
     {
         Schema::create('item_properties', function (Blueprint $table) {
-            $table->integer('id')->unsigned(); // part of composite PK
+            $table->bigIncrements('id')->unsigned(); // part of composite PK
             $table->string('language', 2); // part of composite PK
             $table->foreignId('item_id')->index();
             $table->string('key', 255)->index();
